+++ src/App/index.js
@@ -1,21 +1,22 @@
-import React, { useState } from 'react';
-import WeatherCard from '../Components/WeatherCard';
+import React, { useEffect } from 'react';
+import WeatherForm from '../Components/form';
+// import WeatherCard from '../Components/WeatherCard';
 import getWeather from '../helpers/data/weatherData';
 import './App.scss';
 
 function App() {
-  const [weather, setWeather]= useState({});
-  
+  // const [weather, setWeather]= useState({});
   useEffect(() => {
     getWeather().then((weatherObj) => console.warn(weatherObj));
   }, []);
 
   return (
     <div className='App'>
-      <WeatherCard
+      <WeatherForm />
+      {/* <WeatherCard
       weather={weather}
       setWeather={setWeather}
-    />
+    /> */}
     </div>
   );
 }
