+++ src/Components/form.js
@@ -1,13 +1,16 @@
-import React from 'React';
+import React, { useState } from 'react';
+import {
+  Form,
+  FormGroup,
+  Label,
+  Input,
+  Button
+} from 'reactstrap';
+import getWeather from '../helpers/data/weatherData';
 
-export default function weatherForm() {
-  return (
-    <Form>
-    <FormGroup>
-      <Label for="location">Enter Your Location</Label>
-      <Input type="text" name="location" id="location" placeholder="Location" />
-    </FormGroup>
-    <Button color="info" type="submit">Submit</Button>
-    </Form>
-  );
-}
+export default function WeatherForm() {
+  const [userLocation, setUserLocation] = useState('');
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    g
