+++ src/helpers/data/weatherData.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import firebaseConfig from './apiKeys';
+import firebaseConfig from '../apiKeys';
 
 const weatherKey = firebaseConfig.apiKey;
